\name{computeOutput1}
\alias{computeOutput1}

\title{
Computes output
}
\description{
Computes output (prediction) for a one hidden layer network for one observation
}
\usage{
computeOutput1(x, weights)
}

\arguments{
  \item{x}{
properties of observation
}
  \item{weights}{
weights of the neural network
}
}
\value{
Returns a single numeric value.
}

\author{
Georg Steinbuss
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
