\name{computeOutput2}
\alias{computeOutput2}

\title{
Computes output
}
\description{
Computes output (prediction) for a two hidden layers network for one observation
}
\usage{
computeOutput2(x, weights)
}

\arguments{
  \item{x}{
properties of observation
}
  \item{weights}{
weights of the neural network
}
}
\value{
Returns a single numeric value.
}

\author{
Georg Steinbuss
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line