\name{createWeights1}
\alias{createWeights1}

\title{
Creates random weights
}
\description{
Creates random weights for a single hidden layer network
}
\usage{
createWeights1(I, H)
}

\arguments{
  \item{I}{
number of input neurons
}
  \item{H}{
number of hidden neurons
}
}
\value{
Returns a S4 class object Weights
}

\author{
Georg Steinbuss
}

\seealso{
\link{Weights-class}
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
