\name{crossEntropy}
\alias{crossEntropy}

\title{
Cross entropy
}
\description{
The error function cross entropy
}
\usage{
crossEntropy(x, y)
}

\arguments{
  \item{x}{
properties of observation
}
  \item{y}{
characteristic of observation (zero or one)
}
}

\value{
returns a single numeric value
}

\author{
Georg Steinbuss
}

\seealso{
link{squaredError}
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
