\name{find.Threshold}
\alias{find.Threshold}

\title{
Finds best threshold
}
\description{
Finds the best threshold to transform probabilities in classes
}
\usage{
find.Threshold(obs, stepsize = 0.1, predict)
}

\arguments{
  \item{obs}{
class of observation
}
  \item{stepsize}{
in which step size the threshold is raised
}
  \item{predict}{
prediction of network
}
}

\author{
Georg Steinbuss
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
