\name{is.acct}
\alias{is.acct}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Checks for correct input
}
\description{
Checks for correct input
}
\usage{
is.acct(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
activation function
}
}

\author{
Georg Steinbuss
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
\keyword{ internal }
