\name{is.decay}
\alias{is.decay}

\title{
Checks for correct input
}
\description{
Checks for correct input
}
\usage{
is.decay(x, tol = .Machine$double.eps^0.5)
}

\arguments{
  \item{x}{
decay
}
  \item{tol}{
tolerance
}
}
\author{
Georg Steinbuss
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
\keyword{ internal }
