\name{is.sample}
\alias{is.sample}

\title{
Checks for correct input
}
\description{
Checks for correct input
}
\usage{
is.sample(x)
}

\arguments{
  \item{x}{
Boolean value for sameSample
}
}
\author{
Georg Steinbuss
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
\keyword{ internal }
