\name{is.sampleLeng}
\alias{is.sampleLeng}

\title{
Checks for correct input
}
\description{
Checks for correct input
}
\usage{
is.sampleLeng(x)
}

\arguments{
  \item{x}{
sample length
}
}
\author{
Georg Steinbuss
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
\keyword{ internal }
