\name{is.stepMax}
\alias{is.stepMax}

\title{
Checks for correct input
}
\description{
Checks for correct input
}
\usage{
is.stepMax(x, tol = .Machine$double.eps^0.5)
}

\arguments{
  \item{x}{
maximal number of steps
}
  \item{tol}{
tolerance
}
}
\author{
Georg Steinbuss
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
\keyword{ internal }
