\name{is.thres.error}
\alias{is.thres.error}

\title{
Checks for correct input
}
\description{
Checks for correct input
}
\usage{
is.thres.error(x)
}

\arguments{
  \item{x}{
threshold.TotalError
}
}
\author{
Georg Steinbuss
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
\keyword{ internal }
