\name{logistic.differential}
\alias{logistic.differential}

\title{
Differential of logistic function
}
\description{
Computes value for the differential of the logistic function
}
\usage{
logistic.differential(x)
}

\arguments{
  \item{x}{
Input for differential of logistic function
}
}

\author{
Georg Steinbuss
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
