\name{squaredError}
\alias{squaredError}

\title{
Computes squared error
}
\description{
Computes squared difference between two values
}
\usage{
squaredError(x, y)
}

\arguments{
  \item{x}{
value 1
}
  \item{y}{
value 2
}
}

\author{
Georg Steinbuss
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
