\name{sumSquaredError}
\alias{sumSquaredError}

\title{
Sums up squared error
}
\description{
Computes the full value of the squared error for TeachNet
}
\usage{
sumSquaredError(weights, data, h2)
}

\arguments{
  \item{weights}{
current weights
}
  \item{data}{
data frame
}
  \item{h2}{
number of neurons in second hidden layer
}
}

\author{
Georg Steinbuss
}
\seealso{
\link{crossEntropy}
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
