\name{transformPrediction}
\alias{transformPrediction}

\title{
Transforms prediction
}
\description{
Transforms prediction from prediction to class
}
\usage{
transformPrediction(pred, threshold)
}

\arguments{
  \item{pred}{
Prediction
}
  \item{threshold}{
A threshold up to which a prediction is class 0 or 1. A value from 0 to 1
}
}


\author{
Georg Steinbuss
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
